refactor(commands): Remove duplicate corrections
PostRowFeedback has a unique constraint between user and post_row, because a corrector shouldn't be able to mark a sentence as perfect and be able to correct the same sentence. There were instances where PerfectRow and CorrectedRow both shared the same post_id and user. This is resolved by retaining only the latest correction entry and removing older ones.